requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
libsql-experimental==0.10.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiohttp==3.9.1
websockets==12.0
python-dotenv==1.0.0